feature: [#35] 차단 관련 메서드 BlockParty 패키지로 이동 / 사용자가 차단한 모임 제외하여 조회 및 사용자가 속한 모임 조회 url 변경
